--- src/main/java/com/example/demo/room/dto/RoomGetRoomIdLIstResponse.java
@@ -1,8 +1,0 @@
-package com.example.demo.room.dto;
-
-import lombok.Data;
-
-@Data
-public class RoomGetRoomIdLIstResponse {
-    private String[] roomIdList;
-}
