+++ src/main/java/com/example/demo/user/dto/UserGetRoomIdListResponse.java
@@ -1,0 +1,8 @@
+package com.example.demo.user.dto;
+
+import lombok.Data;
+
+@Data
+public class UserGetRoomIdListResponse {
+    private String[] roomIdList;
+}
